third_party/WebKit/LayoutTests/accessibility/input-type-text-selection-expected.txt

This tests that text selection is reported correctly for text input elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 1
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 2
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 3
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 4
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 0
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 2
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 3
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 4
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 1
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 2
PASS inputAccessible.selectionEnd is 3
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 2
PASS inputAccessible.selectionEnd is 4
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 2
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 2
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 3
PASS inputAccessible.selectionEnd is 4
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 3
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 3
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 4
PASS inputAccessible.selectionEnd is 5
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 4
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS inputAccessible.selectionStart is 5
PASS inputAccessible.selectionEnd is 6
PASS inputAccessible.selectionStartLineNumber is 0
PASS inputAccessible.selectionEndLineNumber is 0
PASS successfullyParsed is true

TEST COMPLETE
